removed redundant test;
